SGD porting and data handling
