+++ func/sgd/svm_begin.cpp
@@ -2,16 +2,51 @@
 #include "faasm/counter.h"
 #include "faasm/sgd.h"
 #include "faasm/matrix.h"
+#include "faasm/func.h"
+#include "faasm/time.h"
+#include "faasm/print.h"
 
 #include <stdio.h>
 
 using namespace faasm;
 
+void epochFinished() {
+    bool fullAsync = getEnvFullAsync();
+    SgdParams p = readParamsFromState(PARAMS_KEY, fullAsync);
+
+    // Get epoch count and barrier count(starts at zero)
+    int thisEpoch = getCounter(EPOCH_COUNT_KEY, p.fullAsync);
+
+    long tsTotal = p.nEpochs * sizeof(double);
+    long tsOffset = thisEpoch * sizeof(double);
+
+    // Record the time for this epoch
+    double ts = faasm::getSecondsSinceEpoch();
+    auto tsBytes = reinterpret_cast<uint8_t *>(&ts);
+    faasmWriteStateOffset(LOSS_TIMESTAMPS_KEY, tsTotal, tsOffset, tsBytes, sizeof(double), p.fullAsync);
+
+    // Record the loss for this epoch
+    double loss = faasm::readRootMeanSquaredError(p);
+    auto lossBytes = reinterpret_cast<uint8_t *>(&loss);
+    faasmWriteStateOffset(LOSSES_KEY, tsTotal, tsOffset, lossBytes, sizeof(double), p.fullAsync);
+
+    // Decay learning rate (apparently hogwild doesn't actually do this although it takes in the param)
+    // p.learningRate = p.learningRate * p.learningDecay;
+    // writeParamsToState(PARAMS_KEY, p);
+
+    // Increment epoch counter
+    incrementCounter(EPOCH_COUNT_KEY, p.fullAsync);
+    int nextEpoch = thisEpoch + 1;
+    char *tsString = faasm::floatToStr(ts);
+    char *lossString = faasm::floatToStr(loss);
+    printf("Starting epoch %i (time = %s, loss = %s)\n", nextEpoch, tsString, lossString);
+}
+
 FAASM_MAIN_FUNC() {
     long inputSize = faasmGetInputSize();
     int nBatches;
     if (inputSize == 0) {
-        nBatches = 200;
+        nBatches = 2;
         printf("SVM running default %i batches \n", nBatches);
     } else {
         auto inputBuffer = new uint8_t[inputSize];
@@ -49,10 +84,104 @@ FAASM_MAIN_FUNC() {
         faasmReadStatePtr(FEATURE_COUNTS_KEY, nFeatureCountBytes, false);
     }
 
-    // Begin first epoch
-    printf("Chaining epoch call\n");
+    // Run the epochs
     initCounter(EPOCH_COUNT_KEY, p.fullAsync);
-    faasmChainFunction("sgd_epoch");
+    for (int i = 0; i < p.nEpochs; i++) {
+        printf("Chaining epoch call\n");
+        int epochCallId = faasmChainThis(1);
+        faasmAwaitCall(epochCallId);
+
+        // Epoch finished at this point
+        epochFinished();
+    }
+
+    return 0;
+}
+
+FAASM_FUNC(epoch, 1) {
+    // Load params
+    bool fullAsync = faasm::getEnvFullAsync();
+    faasm::SgdParams p = faasm::readParamsFromState(PARAMS_KEY, fullAsync);
+
+    // Set per-epoch memory to zero
+    faasm::zeroErrors(p);
+    faasm::zeroFinished(p);
+
+    // Get the epoch count
+    int epoch = faasm::getCounter(EPOCH_COUNT_KEY, p.fullAsync);
+
+    // Shuffle start indices for each batch
+    int *batchNumbers = faasm::randomIntRange(p.nBatches);
+
+    // Chain new calls to perform the work
+    std::vector<int> workerCallIds(p.nBatches);
+    for (int w = 0; w < p.nBatches; w++) {
+        int startIdx = batchNumbers[w] * p.batchSize;
+
+        // Make sure we don't overshoot
+        int endIdx = std::min(startIdx + p.batchSize, p.nTrain - 1);
+
+        // Prepare input data for the worker as a string
+        std::string args = std::to_string(w) + " " + std::to_string(startIdx) + " " + std::to_string(endIdx) + " " +
+                           std::to_string(epoch);
+        printf("Chaining sgd_step with args: %s\n", args.c_str());
+
+        auto inputBytes = reinterpret_cast<const uint8_t *>(args.c_str());
+
+        // Call the chained function
+        int thisCallId = faasmChainThisInput(2, inputBytes, args.size());
+        workerCallIds.push_back(thisCallId);
+    }
+
+    // Wait for all workers to finish
+    for (auto callId : workerCallIds) {
+        faasmAwaitCall(callId);
+    }
+
+    delete[] batchNumbers;
+
+    return 0;
+}
+
+FAASM_FUNC(step, 2) {
+    bool fullAsync = getEnvFullAsync();
+
+    long inputSize = faasmGetInputSize();
+    auto inputBuffer = new uint8_t[inputSize];
+    faasmGetInput(inputBuffer, inputSize);
+
+    const char *inputStr = reinterpret_cast<const char *>(inputBuffer);
+    int *intArgs = faasm::splitStringIntoInts(inputStr, 4);
+
+    int batchNumber = intArgs[0];
+    int startIdx = intArgs[1];
+    int endIdx = intArgs[2];
+    int epoch = intArgs[3];
+
+    printf("SGD step: %i %i %i %i\n", batchNumber, startIdx, endIdx, epoch);
+
+    // Load params
+    SgdParams sgdParams = readParamsFromState(PARAMS_KEY, fullAsync);
+
+    // Perform updates
+    if (sgdParams.lossType == HINGE) {
+        printf("SGD hinge weight update\n");
+        hingeLossWeightUpdate(sgdParams, epoch, batchNumber, startIdx, endIdx);
+    } else {
+        printf("SGD least squares weight update\n");
+        leastSquaresWeightUpdate(sgdParams, batchNumber, startIdx, endIdx);
+    }
+
+    // Flag that this worker has finished
+    printf("Writing finished flag\n");
+    writeFinishedFlag(sgdParams, batchNumber);
+
+    // If this is the last, dispatch the barrier (will have finished by now or will do soon)
+    if (batchNumber == sgdParams.nBatches - 1) {
+        faasmChainFunction("sgd_barrier");
+    }
+
+    delete[] intArgs;
 
     return 0;
 }
