+++ tasks/env.py
@@ -20,6 +20,7 @@ WASM_LIB_DIR = join(PROJ_ROOT, "wasm", "lib")
 RUNTIME_S3_BUCKET = "faasm-runtime"
 STATE_S3_BUCKET = "faasm-state"
 MISC_S3_BUCKET = "faasm-misc"
+DATA_S3_BUCKET = "faasm-data"
 
 # TODO - avoid hard-coding
 AWS_ACCOUNT_ID = "733781933474"
